Moved database dict lookup to its own file
