+++ GEECS-PythonAPI/geecs_python_api/controls/data_acquisition/__init__.py
@@ -1,3 +1,4 @@
 from .data_acquisition import DeviceManager, DataLogger, ActionManager, SoundPlayer
 from .scan_manager import ScanManager
+from .database_dict_lookup import  DatabaseDictLookup
 from .utils import ConsoleLogger, get_full_config_path, visa_config_generator
