+++ GEECS-PythonAPI/geecs_python_api/controls/data_acquisition/scan_manager.py
@@ -16,48 +16,13 @@ import numpy as np
 from nptdms import TdmsWriter, ChannelObject
 
 # Internal project imports
-from geecs_python_api.controls.data_acquisition import DeviceManager, ActionManager, DataLogger
+from geecs_python_api.controls.data_acquisition import DeviceManager, ActionManager, DataLogger, DatabaseDictLookup
 from geecs_python_api.controls.data_acquisition.utils import ConsoleLogger
-from geecs_python_api.controls.interface import load_config, GeecsDatabase
+from geecs_python_api.controls.interface import GeecsDatabase
 from geecs_python_api.controls.interface.geecs_paths_config import GeecsPathsConfig
 from geecs_python_api.controls.devices.geecs_device import GeecsDevice
 from geecs_python_api.analysis.scans.scan_data import ScanData
 from image_analysis.utils import get_imaq_timestamp_from_png, get_picoscopeV2_timestamp, get_custom_imaq_timestamp
-
-
-# TODO rewrote the static code section to be more flexible, but it could use a new home or move to GEECS Python API
-class DatabaseDictLookup:
-    """ Stores the current database dictionary. Reloads upon new Scan Manager instance, but only if the experiment
-     name has changed.  """
-    def __init__(self):
-        self.experiment_name = None
-        self.database_dict = None
-        self.load_config_flag = False
-
-    def get_database(self):
-        return self.database_dict
-
-    def reload(self, experiment_name):
-        if self.experiment_name == experiment_name and self.load_config_flag:
-            return
-        self.load_config_flag = True  # Flag ensures config file is read at least once if no experiment name given
-
-        if experiment_name is None:
-            config = load_config()
-
-            if config and 'Experiment' in config and 'expt' in config['Experiment']:
-                experiment_name = config['Experiment']['expt']
-                logging.info(f"default experiment is: {experiment_name}")
-            else:
-                logging.warning("Configuration file not found or default experiment not defined.")
-
-        self.experiment_name = experiment_name
-        try:
-            GeecsDevice.exp_info = GeecsDatabase.collect_exp_info(experiment_name)
-            self.database_dict = GeecsDevice.exp_info['devices']
-        except AttributeError:
-            logging.warning("Could not retrieve dictionary of GEECS Devices")
-            self.database_dict = None
 
 
 database_dict = DatabaseDictLookup()
